+++ src/Models/CustomerServiceContainer.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace AndrewSvirin\Ebics\Models;
+
+/**
+ * Class CustomerServiceContainer represents CustomerServiceContainer model.
+ *
+ * @license http://www.opensource.org/licenses/mit-license.html  MIT License
+ * @author Serhii Raspopov
+ */
+final class CustomerServiceContainer extends OrderData
+{
+}