+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 * Added CurlHttpClient and PsrHttpClient to use standard client.
 * Updated AbstractX509Generator to handle custom options.
 * Improved Bank-letter
+* Add CustomerServiceContainer for create Service XML-Container (SVC format)